e/requirements-phase1.txt
psycopg2-binary==2.9.9
redis==5.0.8
sentry-sdk==2.8.0
pytest==8.3.2
pytest-flask==1.3.0
requests==2.32.3
backoff==2.2.1
